web/requirements.txt
django>=2.0
django-debug-toolbar==1.7
djangorestframework>=3.7
djangorestframework-gis
django-rest-swagger
drf-extensions
markdown
django-filter
django_extensions

django-cors-headers
datapunt-authorization-django
datapunt-authorization-levels

psycopg2
requests

pytest
pytest-flake8
pytest-django

coreapi

git+https://github.com/Amsterdam/drf_amsterdam.git#egg=drf_amsterdam
